+++ src/app/page.tsx
@@ -4,11 +4,9 @@ import SearchFilter from "~/components/search-filter";
 export default async function Home() {
   return (
     <div className="flex h-[85vh] flex-col">
-      <HeaderLayout>
-        <div className="flex h-full flex-col items-center justify-center">
-          <SearchFilter />
-        </div>
-      </HeaderLayout>
+      <div className="flex h-full flex-col items-center justify-center">
+        <SearchFilter />
+      </div>
     </div>
   );
 }
