+++ src/app/roles/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
+import { Review } from "@prisma/client";
 import { useState } from "react";
-import HeaderLayout from "~/components/header-layout";
 import { ReviewCard } from "~/components/review-card";
 import { ReviewCardPreview } from "~/components/review-card-preview";
 import SearchFilter from "~/components/search-filter";
@@ -11,18 +11,20 @@ import { api } from "~/trpc/react";
 export default function Roles() {
   const reviews = api.review.list.useQuery();
 
-  const [selectedReviewIdx, setSelectedReviewIdx] = useState<number>(0);
+  const [selectedReview, setSelectedReview] = useState<Review | undefined>(
+    reviews.data ? reviews.data[0] : undefined,
+  );
 
   return (
-    <HeaderLayout>
+    <>
       <SearchFilter />
       {/* TODO: Loading animations */}
       {reviews.data && (
         <div className="mb-8 grid w-4/5 grid-cols-5 gap-4 lg:w-3/4">
           <div className="col-span-2 gap-3">
-            {reviews.data.map((review, idx) => {
+            {reviews.data.map((review) => {
               return (
-                <div key={review.id} onClick={() => setSelectedReviewIdx(idx)}>
+                <div key={review.id} onClick={() => setSelectedReview(review)}>
                   <ReviewCardPreview
                     reviewObj={review}
                     className={cn("mb-4 hover:border-2")}
@@ -33,13 +35,11 @@ export default function Roles() {
           </div>
           <div className="col-span-3">
             {reviews.data.length > 0 && reviews.data[0] && (
-              <ReviewCard
-                reviewObj={reviews.data[selectedReviewIdx] || reviews.data[0]}
-              />
+              <ReviewCard reviewObj={selectedReview || reviews.data[0]} />
             )}
           </div>
         </div>
       )}
-    </HeaderLayout>
+    </>
   );
 }
