+++ src/components/header-layout.tsx
@@ -1,5 +1,4 @@
 "use client";
-import { SessionProvider } from "next-auth/react";
 import { ReactNode } from "react";
 import Header from "~/components/header";
 
@@ -11,9 +10,7 @@ import Header from "~/components/header";
 export default function HeaderLayout({ children }: { children: ReactNode }) {
   return (
     <div className="flex min-h-screen flex-col">
-      <SessionProvider>
-        <Header />
-      </SessionProvider>
+      <Header />
       <article className="mt-16 flex w-[100vw] flex-col items-center justify-center gap-16">
         {children}
       </article>
